requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Werkzeug==3.0.3
SQLAlchemy==2.0.30
PyMuPDF==1.24.5
python-docx==1.1.2
scikit-learn==1.5.0
python-dotenv==1.0.1
